Extract main to different class
